PrintWorker: move to helper/model subdirectory
Move also PrintWorkerHandler and printdefs.h
