+++ src/printing/CMakeLists.txt
@@ -3,14 +3,6 @@ add_subdirectory(wizard)
 
 set(MR_TIMETABLE_PLANNER_SOURCES
   ${MR_TIMETABLE_PLANNER_SOURCES}
-  printing/printdefs.h
-
-  printing/printworker.h
-  printing/printworker.cpp
-
-  printing/printworkerhandler.h
-  printing/printworkerhandler.cpp
-
   PARENT_SCOPE
 )
