+++ src/printing/helper/CMakeLists.txt
@@ -3,7 +3,6 @@ add_subdirectory(view)
 
 set(MR_TIMETABLE_PLANNER_SOURCES
   ${MR_TIMETABLE_PLANNER_SOURCES}
-
   PARENT_SCOPE
 )
