+++ src/printing/helper/model/CMakeLists.txt
@@ -7,11 +7,19 @@ set(MR_TIMETABLE_PLANNER_SOURCES
   printing/helper/model/pagesizemodel.h
   printing/helper/model/pagesizemodel.cpp
 
+  printing/helper/model/printdefs.h
+
   printing/helper/model/printhelper.h
   printing/helper/model/printhelper.cpp
 
   printing/helper/model/printpreviewsceneproxy.h
   printing/helper/model/printpreviewsceneproxy.cpp
+
+  printing/helper/model/printworker.h
+  printing/helper/model/printworker.cpp
+
+  printing/helper/model/printworkerhandler.h
+  printing/helper/model/printworkerhandler.cpp
 
   PARENT_SCOPE
 )
