+++ src/printing/wizard/printwizard.h
@@ -4,7 +4,7 @@
 #include <QWizard>
 #include <QVector>
 
-#include "printing/printdefs.h"
+#include "printing/helper/model/printdefs.h"
 
 class QPrinter;
 class SceneSelectionModel;
